+++ src/Tests/Toolchain/JobUninstall.test.ts
@@ -18,22 +18,22 @@ import {assert} from 'chai';
 
 import {Command} from '../../Backend/Command';
 import {Job} from '../../Project/Job';
-import {JobInstall} from '../../Project/JobInstall';
+import {JobUninstall} from '../../Toolchain/JobUninstall';
 
 suite('Project', function() {
-  suite('JobInstall', function() {
+  suite('JobUninstall', function() {
     suite('#contructor()', function() {
       test('is contructed with command', function() {
         let cmd = new Command('apt-get');
-        cmd.push('install');
+        cmd.push('purge');
         cmd.push('pkgname');
 
-        let job = new JobInstall(cmd);
+        let job = new JobUninstall(cmd);
         assert.deepStrictEqual(job.tool, cmd[0]);
         assert.deepStrictEqual(job.toolArgs[0], cmd[1]);
         assert.deepStrictEqual(job.toolArgs[1], cmd[2]);
-        assert.deepStrictEqual(job.name, 'install');
-        assert.equal(job.jobType, Job.Type.tInstall);
+        assert.deepStrictEqual(job.name, 'uninstall');
+        assert.equal(job.jobType, Job.Type.tUninstall);
         assert.isTrue(job.valid);
       });
     });
